fix(macos): Correct FFI signature for CGEventKeyboardSetUnicodeString
CRITICAL FIX for Apple Silicon ARM64 compatibility.

**Problem:**
The FFI declaration used c_uint (32-bit unsigned) for stringLength parameter,
but Apple's API expects CFIndex (c_long, 64-bit signed on 64-bit systems).

**Impact:**
- Stack corruption on Apple Silicon ARM64 due to calling convention mismatch
- Potential crashes when typing text on macOS
- Memory safety violation in FFI boundary

**Fix:**
- Changed stringLength parameter from c_uint to c_long in FFI declaration
- Updated call site to cast utf16.len() as c_long instead of c_uint
- Added documentation comment explaining Apple's actual signature

**Apple's Signature:**
void CGEventKeyboardSetUnicodeString(
    CGEventRef event,
    CFIndex stringLength,     // CFIndex = signed long (64-bit)
    const UniChar *unicodeString
);

This fix is essential for stable operation on macOS before publishing v0.7.1.

+++ rust-crates/swictation-daemon/src/macos_text_inject.rs
@@ -17,18 +17,22 @@ use core_graphics::event::{
     CGEvent, CGEventFlags, CGEventSource, CGEventSourceStateID, CGEventTapLocation, CGEventType,
     CGKeyCode,
 };
-use std::os::raw::{c_uint, c_void};
+use std::os::raw::{c_long, c_uint, c_void};
 use tracing::{debug, warn};
 
 /// FFI declaration for CGEventKeyboardSetUnicodeString
 ///
 /// This function is not exposed by the core-graphics crate, so we declare it manually.
 /// It allows setting Unicode text content for keyboard events.
+///
+/// CRITICAL: stringLength is CFIndex (c_long on 64-bit) NOT c_uint.
+/// Apple's signature: void CGEventKeyboardSetUnicodeString(CGEventRef event, CFIndex stringLength, const UniChar *unicodeString);
+/// where CFIndex = signed long (64-bit on Apple Silicon)
 #[link(name = "CoreGraphics", kind = "framework")]
 extern "C" {
     fn CGEventKeyboardSetUnicodeString(
         event: *mut c_void,
-        stringLength: c_uint,
+        stringLength: c_long,
         unicodeString: *const u16,
     );
 }
@@ -178,7 +182,7 @@ impl MacOSTextInjector {
             unsafe {
                 CGEventKeyboardSetUnicodeString(
                     event.as_ptr() as *mut c_void,
-                    utf16.len() as c_uint,
+                    utf16.len() as c_long,
                     utf16.as_ptr(),
                 );
             }
